feat: cmake 경로 추가

+++ src/Server/db/bank/tables/transaction_logdb.h
@@ -13,8 +13,8 @@ struct TransactionLog
     unsigned int cusId;     // Customer 테이블의 id를 참조. 돈 보낸 사람
     std::string accNum; // 12자리 계좌번호 (조회용)
     unsigned int transaction_type; // 입금시 1, 송금시 2, 출금시 3
-    unsigned int trade_amount;
-    unsigned int remaining_balance;
+    unsigned long long trade_amount;
+    unsigned long long remaining_balance;
     std::string created_at;
 
     static auto getTableDefinition()
